frame: widen parameter type in quic_frame_param_idx()
quic_frame_param_idx() may receive transport parameters that exceed
32-bit range. Update its argument type from u32 to u64 to avoid
truncation and ensure correct indexing for all valid parameter values.

+++ modules/net/quic/frame.c
@@ -3026,7 +3026,7 @@ enum {
 	QUIC_TRANSPORT_PARAM_IDX_MAX,
 };
 
-static u32 quic_frame_param_idx(u32 param)
+static u32 quic_frame_param_idx(u64 param)
 {
 	if (param <= QUIC_TRANSPORT_PARAM_MAX)
 		return param;
